requirements.txt
Django==3.1.3
django-environ==0.4.5
django-modern-rpc==0.12.1
six==1.15.0
vk_api==11.9.0
django-simple-history==2.12.0
python-benedict==0.22.0